refactor(services): move rescore-with-fallback cluster to the services layer
CI's layer-import guard flagged the assistant tools module importing quodeq.api (rescore_with_fallback lived in routes_findings). The whole rescore-after-mutation cluster (slim scores, projection locks, project-all-runs, rescore run + fallback) moves verbatim to services/mutation_rescore.py, where both the findings routes (api -> services, legal) and the dismiss action apply (assistant -> services, legal) import it. The api-layer _project_dir keeps its Flask abort; the service twin raises ValueError. Baseline shrinks by one: the grandfathered api -> data import moved with the cluster into a layer where it is allowed.
